+++ src/Commands/Enums/SearchCrawledPropertySet.cs
@@ -1,0 +1,10 @@
+namespace PnP.PowerShell.Commands.Enums
+{
+    public enum SearchCrawledPropertySet
+    {
+        SharePointDefault,
+        SharePointTaxonomy,
+        SharePointStructured,
+        SharePointRich
+    }
+}